requirements.txt
Django==1.8.18
Pillow==2.8.1
django-debug-toolbar==1.3.0
django-wysiwyg==0.7.0
sqlparse==0.1.15
sorl-thumbnail==12.3
django-constance[database]>=1.1.1
django-versatileimagefield>=1.6
git+https://github.com/CalPolyResDev/django-jfu#egg=django-jfu
django-model-utils>=2.5
unidecode>=0.04.19
django-bootstrap3>=6.2.2
